views/auth/notesharingapp_python/requirements.txt
# FastAPI Stack (Updated)
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pymysql==1.1.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
email-validator==2.1.0
jinja2==3.1.2
aiofiles==23.2.1
starlette==0.36.3
cryptography==41.0.7
